Use two arrays to properly update state
According to the Wikipedia entry explaining Conway's Game of Life, the
rules must be applied simultaneously to all cells; that is, the state of
each cell must be calculated at once. In turn, use two data structures
to maintain the state of the game: one for the current iteration and one
for the next. This way, the state of the game does not change
incrementally. Each iteration, one matrix acts as the reference and only
the other is modified.

Likewise, rearrange files, separating implementation of the 2D array into
a separate unit, and create a build scripts as well.