Simplify state management
Instead of using two 2-dimensional arrays and the default curses window to
maintain the state of the game, use only a single 2-dimensional array
and the curses window.

With this change, the state of each cell no longer needs to be copied to
the curses window each iteration from one of the arrays. Consequently,
it also removes the need for the (unclear) macros previously used to
count a cell's living neighbors.